requirements.txt
copernicusmarine
anemoi-datasets  # Needs some local edits (see README)
xarray
pytest